mcp/bitbucket/prj-000/data-api/requirements.txt
flask==3.0.0
azure-cosmos==4.5.1
pyyaml==6.0.1
